+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/oldConfig/utils/FTCConstants.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.utils;
+package org.firstinspires.ftc.teamcode.oldConfig.utils;
 
 /**
  * This class has all constant definitions.
